+++ src/main/java/ecostruxure/rate/calculator/bll/profile/observers/TeamIProfileAllocationObserver.java
@@ -26,6 +26,5 @@ public class TeamIProfileAllocationObserver implements IProfileObserver {
                     .divide(new BigDecimal("100.00"), 2, BigDecimal.ROUND_HALF_UP)));
             teamProfileRepository.save(teamProfile);
         }
-
     }
 }
