--- src/main/java/ecostruxure/rate/calculator/bll/service/HistoryService.java
@@ -1,19 +1,0 @@
-//package ecostruxure.rate.calculator.bll.service;
-//
-//import java.util.List;
-//import java.util.UUID;
-//
-//public class HistoryService {
-//    private final IHistoryDAO historyDAO;
-//    public HistoryService() throws Exception {
-//        this.historyDAO = new HistoryDAO();
-//    }
-//
-//    public List<ProfileHistory> getProfileHistory(UUID profileId) throws Exception {
-//        return historyDAO.getProfileHistory(profileId);
-//    }
-//
-//    public List<TeamHistory> getTeamHistory(UUID teamId) throws Exception {
-//        return historyDAO.getTeamHistory(teamId);
-//    }
-//}
