+++ src/main/java/ecostruxure/rate/calculator/controllers/ProjectController.java
@@ -2,7 +2,7 @@ package ecostruxure.rate.calculator.controllers;
 
 import ecostruxure.rate.calculator.be.Project;
 import ecostruxure.rate.calculator.be.dto.ProjectDTO;
-import ecostruxure.rate.calculator.bll.service.ProjectService;
+import ecostruxure.rate.calculator.bll.project.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
